+++ packages/common/src/types.ts
@@ -1575,6 +1575,12 @@ export interface IdObjectRequestV1 {
     proofsOfKnowledge: string;
 }
 
+export interface IdRecoveryRequest {
+    idCredPub: string;
+    timestamp: bigint;
+    proof: string;
+}
+
 export interface AttributeList {
     validTo: string; // "YYYYMM"
     createdAt: string; // "YYYYMM"