+++ packages/web/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Bumped @concordium/common-sdk to 3.0.0. (Which includes breaking changes to schema versioning)
 
-## 0.4.0 2022-8-8
+## 0.4.0 2022-8-15
 
 ### Changed